Move postgreSQL implement to sharding-scaling-postgresql module
