+++ sharding-scaling-mysql/src/test/java/info/avalon566/shardingscaling/mysql/binlog/codec/MySQLLengthFieldBasedFrameEncoderTest.java
@@ -30,7 +30,6 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-
 @RunWith(MockitoJUnitRunner.class)
 public class MySQLLengthFieldBasedFrameEncoderTest {
