+++ sharding-scaling-mysql/src/test/java/info/avalon566/shardingscaling/mysql/binlog/packet/binlog/RowsEventTest.java
@@ -38,7 +38,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-
 public class RowsEventTest {
 
     private BinlogEventHeader binlogEventHeader;
