+++ sharding-scaling-postgresql/src/main/java/info/avalon556/shardingscaling/postgresql/PostgreSQLJdbcReader.java
@@ -15,36 +15,19 @@
  * limitations under the License.
  */
 
-package info.avalon566.shardingscaling.core.sync.postgresql;
+package info.avalon556.shardingscaling.postgresql;
 
-import info.avalon566.shardingscaling.core.sync.AbstractRunner;
-import info.avalon566.shardingscaling.core.sync.channel.Channel;
-import info.avalon566.shardingscaling.core.sync.reader.LogReader;
-import lombok.Setter;
+import info.avalon566.shardingscaling.core.config.RdbmsConfiguration;
+import info.avalon566.shardingscaling.core.sync.reader.AbstractJdbcReader;
 
 /**
- * PostgreSQL WAL reader.
+ * PostgreSQL JDBC reader.
  *
  * @author avalon566
  */
-public final class PostgreSQLWalReader extends AbstractRunner implements LogReader {
+public final class PostgreSQLJdbcReader extends AbstractJdbcReader {
 
-    @Setter
-    private Channel channel;
-
-    @Override
-    public void run() {
-        //TODO
-        read(channel);
-    }
-
-    @Override
-    public void read(final Channel channel) {
-        throw new UnsupportedOperationException();
-    }
-    
-    @Override
-    public void markPosition() {
+    public PostgreSQLJdbcReader(final RdbmsConfiguration rdbmsConfiguration) {
+        super(rdbmsConfiguration);
     }
 }
-
