+++ sharding-scaling-postgresql/src/main/java/info/avalon556/shardingscaling/postgresql/PostgreSQLWriter.java
@@ -15,19 +15,20 @@
  * limitations under the License.
  */
 
-package info.avalon566.shardingscaling.core.sync.postgresql;
+package info.avalon556.shardingscaling.postgresql;
 
 import info.avalon566.shardingscaling.core.config.RdbmsConfiguration;
-import info.avalon566.shardingscaling.core.sync.reader.AbstractJdbcReader;
+import info.avalon566.shardingscaling.core.sync.writer.AbstractJdbcWriter;
 
 /**
- * PostgreSQL JDBC reader.
+ * postgreSQL writer.
  *
  * @author avalon566
  */
-public final class PostgreSQLJdbcReader extends AbstractJdbcReader {
+public class PostgreSQLWriter extends AbstractJdbcWriter {
 
-    public PostgreSQLJdbcReader(final RdbmsConfiguration rdbmsConfiguration) {
+    public PostgreSQLWriter(final RdbmsConfiguration rdbmsConfiguration) {
         super(rdbmsConfiguration);
     }
 }
+
