+++ sharding-scaling-postgresql/src/main/resources/services/info.avalon566.shardingscaling.core.spi.ScalingEntry
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-info.avalon566.shardingscaling.core.sync.postgresql.PostgreSQLScalingEntry
+info.avalon556.shardingscaling.postgresql.PostgreSQLScalingEntry
